corrected issue with App rerendering
